Voting booth now has better integrity checking when removing a poll.

+++ src/specials/Specs/object_specs/voting_booth.spec
@@ -553,7 +553,7 @@ void
 voting_booth_forget(int acct_id)
 {
 	struct voting_poll *cur_poll;
-	struct memory_rec_struct *cur_mem, *doomed_mem;
+	struct memory_rec_struct *cur_mem, *doomed_mem = NULL;
 	
 	// Remove account from all voting polls
 	for (cur_poll = voting_poll_list; cur_poll; cur_poll = cur_poll->next) {
@@ -571,6 +571,10 @@ voting_booth_forget(int acct_id)
 		}
 	}
 
+	if (!doomed_mem) {
+		errlog("Can't happen at %s:%d", __FILE__, __LINE__);
+		return;
+	}
 	free(doomed_mem);
 	sql_exec("delete from voting_accounts where account=%d", acct_id);
 }
